[#78] feat: add AudioRenderingService
